+++ sync-for-payables/src/codat_sync_for_payables/models/operations/create_bill.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 from codat_sync_for_payables.models.shared import billprototype as shared_billprototype
-from codat_sync_for_payables.types import BaseModel
+from codat_sync_for_payables.types import BaseModel, UNSET_SENTINEL
 from codat_sync_for_payables.utils import (
     FieldMetadata,
     HeaderMetadata,
@@ -10,6 +10,7 @@ from codat_sync_for_payables.utils import (
     RequestMetadata,
 )
 import pydantic
+from pydantic import model_serializer
 from typing import Optional
 from typing_extensions import Annotated, NotRequired, TypedDict
 
@@ -50,3 +51,19 @@ class CreateBillRequest(BaseModel):
         Optional[shared_billprototype.BillPrototype],
         FieldMetadata(request=RequestMetadata(media_type="application/json")),
     ] = None
+
+    @model_serializer(mode="wrap")
+    def serialize_model(self, handler):
+        optional_fields = set(["Idempotency-Key", "billPrototype"])
+        serialized = handler(self)
+        m = {}
+
+        for n, f in type(self).model_fields.items():
+            k = f.alias or n
+            val = serialized.get(k)
+
+            if val != UNSET_SENTINEL:
+                if val is not None or k not in optional_fields:
+                    m[k] = val
+
+        return m