+++ sync-for-payables/src/codat_sync_for_payables/models/operations/create_supplier.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from codat_sync_for_payables.models.shared import (
     supplierprototype as shared_supplierprototype,
 )
-from codat_sync_for_payables.types import BaseModel
+from codat_sync_for_payables.types import BaseModel, UNSET_SENTINEL
 from codat_sync_for_payables.utils import (
     FieldMetadata,
     HeaderMetadata,
@@ -12,6 +12,7 @@ from codat_sync_for_payables.utils import (
     RequestMetadata,
 )
 import pydantic
+from pydantic import model_serializer
 from typing import Optional
 from typing_extensions import Annotated, NotRequired, TypedDict
 
@@ -52,3 +53,19 @@ class CreateSupplierRequest(BaseModel):
         Optional[shared_supplierprototype.SupplierPrototype],
         FieldMetadata(request=RequestMetadata(media_type="application/json")),
     ] = None
+
+    @model_serializer(mode="wrap")
+    def serialize_model(self, handler):
+        optional_fields = set(["Idempotency-Key", "supplierPrototype"])
+        serialized = handler(self)
+        m = {}
+
+        for n, f in type(self).model_fields.items():
+            k = f.alias or n
+            val = serialized.get(k)
+
+            if val != UNSET_SENTINEL:
+                if val is not None or k not in optional_fields:
+                    m[k] = val
+
+        return m