+++ sync-for-payables/src/codat_sync_for_payables/models/shared/bankaccountmappingoption.py
@@ -97,46 +97,39 @@ class BankAccountMappingOption(BaseModel):
 
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
-        optional_fields = [
-            "id",
-            "name",
-            "accountNumber",
-            "nominalCode",
-            "sortCode",
-            "currency",
-            "status",
-            "accountType",
-            "sourceModifiedDate",
-        ]
-        nullable_fields = [
-            "name",
-            "accountNumber",
-            "nominalCode",
-            "sortCode",
-            "currency",
-        ]
-        null_default_fields = []
-
+        optional_fields = set(
+            [
+                "id",
+                "name",
+                "accountNumber",
+                "nominalCode",
+                "sortCode",
+                "currency",
+                "status",
+                "accountType",
+                "sourceModifiedDate",
+            ]
+        )
+        nullable_fields = set(
+            ["name", "accountNumber", "nominalCode", "sortCode", "currency"]
+        )
         serialized = handler(self)
-
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
-            serialized.pop(k, None)
+            is_nullable_and_explicitly_set = (
+                k in nullable_fields
+                and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
+            )
 
-            optional_nullable = k in optional_fields and k in nullable_fields
-            is_set = (
-                self.__pydantic_fields_set__.intersection({n})
-                or k in null_default_fields
-            )  # pylint: disable=no-member
-
-            if val is not None and val != UNSET_SENTINEL:
-                m[k] = val
-            elif val != UNSET_SENTINEL and (
-                not k in optional_fields or (optional_nullable and is_set)
-            ):
-                m[k] = val
+            if val != UNSET_SENTINEL:
+                if (
+                    val is not None
+                    or k not in optional_fields
+                    or is_nullable_and_explicitly_set
+                ):
+                    m[k] = val
 
         return m